back/requirements.txt
fastapi
uvicorn
python-jose
sqlalchemy
argon2-cffi
groq
